Add files via upload
This file contains the first two steps of building the calculator through TOP. I have added the basic functions for adding, subtracting, multiplying, and dividing. I also added the operate function that takes an operator as well as two numbers and calls the above mentioned function above on the numbers.