Add files via upload
In this commit, I have added just a one more from following TOP, step 3. In this step I got stuck when my clear button did not work and gave an error message of null. I learned to put the script tag at the bottom of the html page so that the html renders before the javascript so that the javascript can read the properties correctly. I have added all the buttons on an HTML page and made a clear button. I also added a css page but have not started styling yet.